+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 63.4.0
 
 * Add `stub_any_publishing_api_unreserve_path` test helper.
 * Fix Publishing API `stub_any_publishing_api_call` methods only operating on